requirements.txt
djangorestframework_jwt==1.11.0
drf_yasg==1.17.0
PyMySQL==0.9.3
six==1.14.0
django_filter==2.0.0
Django==3.0.8
PyJWT==1.7.1
